Move `_float_or_none` and `_verify_value_type` helper to util
